responsible-ai-fairness/responsible-ai-fairness/requirements/requirements.txt
fastapi==0.115.11
pydantic==2.9.2 
uvicorn==0.32.0 
aif360==0.6.1
pandas==2.1.1
pyyaml==6.0.2
spacy==3.6.1
python-multipart==0.0.18
BlackBoxAuditing== 0.1.54
pymongo==4.10.1
python-dotenv==1.0.1
joblib==1.4.2
fairlearn==0.10.0
pyarrow==19.0.1
openai==1.52.2
azure.identity==1.19.0
google-generativeai==0.8.3
transformers==4.48.1
python-jose== 3.4.0
torch==2.6.0
holisticai == 0.7.3 
numpy == 1.26.4
scipy==1.11.2
fpdf==1.7.2
backoff==2.2.1
seaborn==0.13.2
httpx==0.27.0
cvxpy==1.6.0
matplotlib==3.9.4
boto3
langchain-aws
../lib/infosys_responsible_ai_fairness-1.1.5-py2.py3-none-any.whl
